Refactor: Generalize component into ToggleInput
